+++ src/main/java/io/wiretap/configuration/WiretapAppLogConfiguration.java
@@ -13,7 +13,7 @@ import org.springframework.context.annotation.Configuration;
 import java.util.List;
 
 @Configuration
-@EnableConfigurationProperties(WiretapAppLogProperties.class)
+@EnableConfigurationProperties({WiretapAppLogProperties.class, WiretapAsyncLoggingProperties.class})
 public class WiretapAppLogConfiguration {
 
     @Autowired